+++ artevolver-core/src/main/java/com/rndmodgames/evolver/PalleteColor.java
@@ -4,11 +4,11 @@ import java.awt.Color;
 
 public class PalleteColor {
 
-	Palette pallete;
+	private Palette pallete;
 
-	Long id;
-	String name;
-	Color color;
+	private Long id;
+	private String name;
+	private Color color;
 
 	public PalleteColor(Palette pallete, Long id, String name, int r, int g, int b){
 
@@ -17,14 +17,6 @@ public class PalleteColor {
 		this.name = name;
 		this.color = new Color(r, g, b);
 	}
-	
-//	public PalleteColor(Palette pallete, Long id, String name, int r, int g, int b, int alpha){
-//
-//        this.pallete = pallete;
-//        this.id = id;
-//        this.name = name;
-//        this.color = new Color(r, g, b, alpha);
-//    }
 
 	public Color getColor() {
 		return color;
@@ -32,5 +24,34 @@ public class PalleteColor {
 
 	public void setColor(Color color) {
 		this.color = color;
+	}
+
+	public Palette getPallete() {
+		return pallete;
+	}
+
+	public void setPallete(Palette pallete) {
+		this.pallete = pallete;
+	}
+
+	public Long getId() {
+		return id;
+	}
+
+	public void setId(Long id) {
+		this.id = id;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public void setName(String name) {
+		this.name = name;
+	}
+
+	@Override
+	public String toString() {
+		return name + " [" + id + "] (" + color.getRed() + "," + color.getGreen() + "," + color.getBlue() + ")";
 	}
 }