+++ artevolver-core/src/main/java/com/rndmodgames/evolver/Triangle.java
@@ -10,14 +10,24 @@ public class Triangle extends Polygon{
 	
 	private Long colorId;
 	private Color color;
+	private PalleteColor palleteColor;
 	
 	private int [] xPoly;
 	private int [] yPoly;
 	private int lenght;
-	
+
 	public Triangle(int [] xPoly, int [] yPoly, int lenght, Color color){
 		super(xPoly, yPoly, lenght);
 		this.color = color;
+		this.xPoly = xPoly;
+		this.yPoly = yPoly;
+		this.lenght = lenght;
+	}
+
+	public Triangle(int [] xPoly, int [] yPoly, int lenght, Color color, PalleteColor palleteColor){
+		super(xPoly, yPoly, lenght);
+		this.color = color;
+		this.palleteColor = palleteColor;
 		this.xPoly = xPoly;
 		this.yPoly = yPoly;
 		this.lenght = lenght;
@@ -33,12 +43,31 @@ public class Triangle extends Polygon{
 		this.lenght = lenght;
 	}
 
+	public Triangle(int[] xPoly, int[] yPoly, int lenght, Long colorId, Color color, PalleteColor palleteColor) {
+		super(xPoly, yPoly, lenght);
+		
+		this.colorId = colorId;
+		this.color = color;
+		this.palleteColor = palleteColor;
+		this.xPoly = xPoly;
+		this.yPoly = yPoly;
+		this.lenght = lenght;
+	}
+
 	public Color getColor() {
 		return color;
 	}
 
 	public void setColor(Color color) {
 		this.color = color;
+	}
+
+	public PalleteColor getPalleteColor() {
+		return palleteColor;
+	}
+
+	public void setPalleteColor(PalleteColor palleteColor) {
+		this.palleteColor = palleteColor;
 	}
 
 	public int[] getxPoly() {